Use unsigned data type

+++ rgb24_15.h
@@ -25,11 +25,11 @@
  * For more information, please refer to <http://unlicense.org>
 */
 
-#define RGB24_TO_15(x) ((((x & 0x0000FF) >> 3) & 0x001F) | \
+#define RGB24_TO_15(x) (unsigned short)((((x & 0x0000FF) >> 3) & 0x001F) | \
 		(((x & 0x00FF00) >> 6) & 0x03E0) | \
 		(((x & 0xFF0000) >> 9) & 0x7C00))
 
-short rgb24_to_15(long int rgb24)
+unsigned short rgb24_to_15(unsigned long int rgb24)
 {
 	return RGB24_TO_15(rgb24); 
 }
